📦 prefer npm specifiers over jsr in import map
Switch @standard-schema/spec to npm, remove unused deps, and document
the npm preference in AGENTS.md.
