+++ deno.json
@@ -4,13 +4,10 @@
   "license": "MIT",
   "publish": { "include": ["vendor", "lib", "mod.ts", "README.md"] },
   "imports": {
-    "@standard-schema/spec": "jsr:@standard-schema/spec@^1.0.0",
-    "@std/assert": "jsr:@std/assert@1",
-    "@std/cli": "jsr:@std/cli@^1.0.21",
+    "@standard-schema/spec": "npm:@standard-schema/spec@^1.0.0",
     "@std/expect": "jsr:@std/expect@^1.0.16",
     "@std/testing": "jsr:@std/testing@^1.0.15",
     "arktype": "npm:arktype@^2.1.20",
-    "sprintf-js": "npm:sprintf-js@^1.1.3",
     "ts-case-convert": "npm:ts-case-convert@^2.1.0",
     "dnt": "jsr:@deno/dnt@0.42.3"
   },
