+++ examples/commands.ts
@@ -1,4 +1,4 @@
-import { assert } from "@std/assert/assert";
+import assert from "node:assert";
 import { type } from "arktype";
 import { cli, field } from "../lib/field.ts";
 import { object } from "../lib/object.ts";
