+++ lib/mod.ts
@@ -3,7 +3,6 @@ export * from "./constant.ts";
 export * from "./field.ts";
 export * from "./object.ts";
 export * from "./commands.ts";
-export * from "./parse.ts";
 export * from "./help.ts";
 export * from "./program.ts";
 export * from "./sequence.ts";
